Fix gcc-6.2 build issues
I normally use clang because the generated executable runs faster (in my
setup); but gcc 6.2.0 apparently catches more warnings and flat-out errors.
Fixed.

Bumped version to "1.2-release".

+++ TransTable.cpp
@@ -138,7 +138,7 @@ size_t TransTable::sanitizeSize(int64 size)
 {
     return
         size < -1 ? 0 : // bad parameter
-        MIN((uint64) normalizeSize(size), normalizeSize(MaxSize()));
+        MIN(normalizeSize(size), normalizeSize(MaxSize()));
 }
 
 void TransTable::resetEntries()
